First time internal waves are witnessed. Boussinesq force is implemented is Guo's style but not necesseray.
What matteers is that the temperature is O(1) so that it is felt by the populations. The value of of N2 in the initial temp gradient can be reduce through the mach number (changing dt). But one has to be careful to rescale g accordingly !
Now the code does that.